+++ src/main/java/com/themythichunter/fireballwand/FireballWand.java
@@ -3,8 +3,12 @@ package com.themythichunter.fireballwand;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public final class FireballWand extends JavaPlugin {
+    private final ConfigHandler configHandler = new ConfigHandler(this);
+
     @Override
     public void onEnable() {
+        configHandler.loadConfig();
+
         getServer().getPluginManager().registerEvents(new FireBallWandListeners(this), this);
         getLogger().info("FireballWand has been enabled!");
     }