+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-11-10
+
+### Added
+
+- Customizable using a `config.yml` for the following settings:
+    - Cooldown duration (default 5000ms).
+    - Blaze Rod custom model data (default 1000727).
+
 ## [1.0.0] - 2023-11-10
 
 ### Added